// backend/src/auth/role.enum.ts
export enum UserRole {
  ULTIMATE_ADMIN = 'ULTIMATE_ADMIN',
  SUPERADMIN = 'SUPERADMIN',
  SENIOR_CHIEF_SECURITY_ADMIN = 'SENIOR_CHIEF_SECURITY_ADMIN',
  CHIEF_DEVELOPMENT_ADMIN = 'CHIEF_DEVELOPMENT_ADMIN',
  CHIEF_SECURITY_ADMIN = 'CHIEF_SECURITY_ADMIN',
  VICE_CHIEF_SECURITY_ADMIN = 'VICE_CHIEF_SECURITY_ADMIN',
  SENIOR_ADMIN = 'SENIOR_ADMIN',
  ADMIN = 'ADMIN',
  JUNIOR_ADMIN = 'JUNIOR_ADMIN',
  EMPLOYEE = 'EMPLOYEE',
  USER = 'USER',
}

export const ROLE_HIERARCHY: Record<UserRole, number> = {
  [UserRole.ULTIMATE_ADMIN]: 100,
  [UserRole.SUPERADMIN]: 90,
  [UserRole.SENIOR_CHIEF_SECURITY_ADMIN]: 80,
  [UserRole.CHIEF_DEVELOPMENT_ADMIN]: 70,
  [UserRole.CHIEF_SECURITY_ADMIN]: 60,
  [UserRole.VICE_CHIEF_SECURITY_ADMIN]: 50,
  [UserRole.SENIOR_ADMIN]: 40,
  [UserRole.ADMIN]: 35,
  [UserRole.JUNIOR_ADMIN]: 30,
  [UserRole.EMPLOYEE]: 20,
  [UserRole.USER]: 10,
};